--- src/main/java/UsedStore/Vo/MyPageVO.java
@@ -1,19 +1,0 @@
-package UsedStore.Vo;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class MyPageVO {
-    private int itemID;
-    private String userID;
-    private String itemName;
-    private String itemContent;
-    private String itemPrice;
-    private String categoryId;
-    private String itemGender;
-    private String itemTag;
-    private String itemTitle;
-    private int status;
-}
