--- src/main/resources/mapper/myPageMapper.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-
-<mapper namespace="mapper.myPageMapper">
-    <select id="getPrice" parameterType="java.lang.String" resultType="map">
-        Select * FROM item WHERE user_id = #{userId}
-    </select>
-</mapper>
